Typed Arrays implemented. Fix bugs with scope capturing, link variables and caching. Improve performance.

+++ NiL.JS/Core/BaseTypes/Uint8Array.cs
@@ -8,7 +8,7 @@ using NiL.JS.Core.Modules;
 namespace NiL.JS.Core.BaseTypes
 {
     [Serializable]
-    public sealed class Int8Array : TypedArray
+    public sealed class Uint8Array : TypedArray
     {
         protected override JSObject this[int index]
         {
@@ -27,9 +27,9 @@ namespace NiL.JS.Core.BaseTypes
             }
         }
 
-        private sbyte getValue(int index)
+        private byte getValue(int index)
         {
-            return (sbyte)buffer.Data[index + byteOffset];
+            return buffer.Data[index + byteOffset];
         }
 
         public override int BYTES_PER_ELEMENT
@@ -37,50 +37,50 @@ namespace NiL.JS.Core.BaseTypes
             get { return 1; }
         }
 
-        public Int8Array()
+        public Uint8Array()
             : base()
         {
         }
 
-        public Int8Array(int length)
+        public Uint8Array(int length)
             : base(length)
         {
         }
 
-        public Int8Array(ArrayBuffer buffer)
+        public Uint8Array(ArrayBuffer buffer)
             : base(buffer, 0, buffer.byteLength)
         {
         }
 
-        public Int8Array(ArrayBuffer buffer, int bytesOffset)
+        public Uint8Array(ArrayBuffer buffer, int bytesOffset)
             : base(buffer, bytesOffset, buffer.byteLength - bytesOffset)
         {
         }
 
-        public Int8Array(ArrayBuffer buffer, int bytesOffset, int length)
+        public Uint8Array(ArrayBuffer buffer, int bytesOffset, int length)
             : base(buffer, bytesOffset, length)
         {
         }
 
-        public Int8Array(JSObject src)
+        public Uint8Array(JSObject src)
             : base(src) { }
 
         [ParametersCount(2)]
         public override TypedArray subarray(Arguments args)
         {
-            return subarrayImpl<Int8Array>(args[0], args[1]);
+            return subarrayImpl<Uint8Array>(args[0], args[1]);
         }
 
         [Hidden]
         public override Type ElementType
         {
             [Hidden]
-            get { return typeof(sbyte); }
+            get { return typeof(byte); }
         }
 
         protected internal override System.Array ToNativeArray()
         {
-            var res = new sbyte[length.iValue];
+            var res = new byte[length.iValue];
             for (var i = 0; i < res.Length; i++)
                 res[i] = getValue(i);
             return res;
